api/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
openai==1.35.13